usb: gadget: f_ipc: Update online flag in set_alt/disable
Currently the driver updates online flag to true as a part of
connect function_work handling. If the function suspend is issued
by the host before this connect work could get scheduled, then
the online flag will be set to true but ep_queue will fail and
it will wait for function resume to get called. While this is
functionally not an issue, it is logically incorrect.
Fix this by moving the updation of online flag out of workqueue
to set_alt/disable functions.

Change-Id: I633b33cc50bfaac1f7281e3ab28fcbbdf9ec72e0
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>