Makefile: Fix device not booting with GCC 7.x and above
@nathanchance: Filtered from the original commit
(zanezam/boeffla-kernel-cm-n51x0@228cc26)
I tested all of the flags in the above commit and this is
the only# one that prevented the phone from booting, with
-Os, -O2, and -O3.# Changes to be committed:
Without this, the phone will boot to the Google logo,
hang for a# few seconds, then reboot, and repeat the
process. This flag can be read about below. This issue
can also be resolved (supposedly) with the -mstrict-align
flag, leading me to believe it is something with memory
access. I believe this is a better solution though given
that ARM handles unaligned access perfectly fine.
https://gcc.gnu.org/onlinedocs/gcc/Optimize-Options.html#index-fstore-merging
https://gcc.gnu.org/onlinedocs/gcc/AArch64-Options.html#index-mstrict-align
Signed-off-by: Nathan Chancellor
<natechancellor@gmail.com>
(cherry picked from commit b791f3c48d534ed82c986170861f165bcabb8820